requirements.txt
langchain
langchain-google-genai
langchain-huggingface
chromadb
ragas
fastapi
uvicorn
datasets
python-dotenv
pydantic
